Extract getVariableType
