+++ plugins/typescript/src/generators/generateReactQueryComponents.ts
@@ -15,12 +15,15 @@ import {
   SchemaObject,
 } from "openapi3-ts";
 import { get, groupBy, uniqBy } from "lodash";
+
 import {
   getType,
   schemaToTypeAliasDeclaration,
 } from "../core/schemaToTypeAliasDeclaration";
 import { findCompatibleMediaType } from "../core/findCompatibleMediaType";
 import { getUsedImports } from "../core/getUsedImports";
+import { getVariablesType } from "../core/getVariablesType";
+
 import { getCustomFetcher } from "../templates/customFetcher";
 import { getContext } from "../templates/context";
 
@@ -549,101 +552,6 @@ const getRequestBodyType = ({
 };
 
 /**
- * Get fetcher variables types from the operation types
- */
-const getVariablesType = ({
-  requestBodyType,
-  requestBodyOptional,
-  headersType,
-  headersOptional,
-  pathParamsType,
-  pathParamsOptional,
-  queryParamsType,
-  queryParamsOptional,
-  contextTypeName,
-}: {
-  requestBodyType: ts.TypeNode;
-  requestBodyOptional: boolean;
-  headersType: ts.TypeNode;
-  headersOptional: boolean;
-  pathParamsType: ts.TypeNode;
-  pathParamsOptional: boolean;
-  queryParamsType: ts.TypeNode;
-  queryParamsOptional: boolean;
-  contextTypeName: string;
-}) => {
-  const variablesItems: ts.TypeElement[] = [];
-
-  const hasProperties = (node: ts.Node) => {
-    return (
-      (!ts.isTypeLiteralNode(node) || node.members.length > 0) &&
-      node.kind !== ts.SyntaxKind.UndefinedKeyword
-    );
-  };
-
-  if (hasProperties(requestBodyType)) {
-    variablesItems.push(
-      f.createPropertySignature(
-        undefined,
-        f.createIdentifier("body"),
-        requestBodyOptional
-          ? f.createToken(ts.SyntaxKind.QuestionToken)
-          : undefined,
-        requestBodyType
-      )
-    );
-  }
-  if (hasProperties(headersType)) {
-    variablesItems.push(
-      f.createPropertySignature(
-        undefined,
-        f.createIdentifier("headers"),
-        headersOptional
-          ? f.createToken(ts.SyntaxKind.QuestionToken)
-          : undefined,
-        headersType
-      )
-    );
-  }
-  if (hasProperties(pathParamsType)) {
-    variablesItems.push(
-      f.createPropertySignature(
-        undefined,
-        f.createIdentifier("pathParams"),
-        pathParamsOptional
-          ? f.createToken(ts.SyntaxKind.QuestionToken)
-          : undefined,
-        pathParamsType
-      )
-    );
-  }
-  if (hasProperties(queryParamsType)) {
-    variablesItems.push(
-      f.createPropertySignature(
-        undefined,
-        f.createIdentifier("queryParams"),
-        queryParamsOptional
-          ? f.createToken(ts.SyntaxKind.QuestionToken)
-          : undefined,
-        queryParamsType
-      )
-    );
-  }
-
-  const fetcherOptionsType = f.createIndexedAccessTypeNode(
-    f.createTypeReferenceNode(f.createIdentifier(contextTypeName), undefined),
-    f.createLiteralTypeNode(f.createStringLiteral("fetcherOptions"))
-  );
-
-  return variablesItems.length === 0
-    ? fetcherOptionsType
-    : f.createIntersectionTypeNode([
-        f.createTypeLiteralNode(variablesItems),
-        fetcherOptionsType,
-      ]);
-};
-
-/**
  * Create the declaration of the fetcher function.
  *
  * @returns Array of nodes
@@ -1155,16 +1063,3 @@ const shouldExtractNode = (node: ts.Node) =>
  */
 const camelizedPathParams = (url: string) =>
   url.replace(/\{\w*\}/g, (match) => `{${c.camel(match)}}`);
-
-/**
- * Returns `true` if all properties are marked as optional.
- *
- * @param node
- * @returns
- */
-const hasAllPropertiesOptional = (node: ts.Node) => {
-  if (!ts.isTypeLiteralNode(node)) return true;
-  return node.members.reduce((mem, member) => {
-    return mem && Boolean(member.questionToken);
-  }, true);
-};
